harden the store scan and narrow the frontmatter guardrail
Fresh-eyes review of the first two slices found two ways a single bad file could
take out the whole command, and one way the guardrail could take out a whole repo.

- an unreadable file or directory no longer aborts the scan: it becomes a node
  issue or a store scan_issue, so doctor can report the thing it exists to find
- symlinked directories are not descended into (entry.file_type, not path.is_dir),
  with a 32-level cap on real nesting; a self-referential link previously recursed
  until the OS refused
- EAGERNESS_KEYS drops path/context/store/scope/root: those are common in foreign
  frontmatter (ticket templates, skill manifests) and would hard-fail doctor on
  conventions rata does not own. Unknown-key warnings now fire only on a near-miss
  of a schema key, which is typo detection instead of policing
- tier 2 rejects label fragments and ordered lists, so a real H1 wins over
  "Original:" and "Phase: build Updated: ..."
- tags: with a shape rata cannot read is reported, not silently dropped
- shadowed store layers are recorded on the node instead of discarded
- --depth 0 is rejected rather than returning nothing; .MD matches; refs containing
  : or # are reported as unaddressable
- test temp_dir helpers consolidated; tests added for each of the above

Verified: 42 tests pass, clippy clean for touched code, zero warnings across the
five real stores, and the tickets outline now reads as real titles.

+++ src/cli.rs
@@ -83,8 +83,8 @@ pub enum Commands {
         /// Limit the outline to a single store. Defaults to every resolved store.
         store: Option<String>,
         /// Cap how many path segments deep into a store nodes are listed. Defaults to unlimited.
-        #[arg(long)]
-        depth: Option<usize>,
+        #[arg(long, value_parser = clap::value_parser!(u16).range(1..))]
+        depth: Option<u16>,
         /// Resolve relative to this directory instead of the current working directory.
         #[arg(long)]
         cwd: Option<PathBuf>,
